Adopted the two header file approach, removed position_iterator code
